==> packages/wmb/wmb-0.1.32.tar.gz/wmb-0.1.32/analysis/integration/mc_atac/Template/02.prepare_atac.ipynb <==
## Import

from ALLCools.clustering import *
from ALLCools.mcds import MCDS
from wmb import brain, cemba_atac, mm10

import pandas as pd
import anndata

import matplotlib.pyplot as plt
from ALLCools.plot import *

n_cell = 5
zscore_abs_cutoff = 3

## Select cells

cells = pd.read_csv('atac_cells.txt', index_col=0, header=None).index
cells.name = 'cell'
cells.size

## Get adata with basic feature selection

atac_ds = MCDS.open(cemba_atac.CEMBA_ATAC_ZARR_PATH,
                    use_obs=cells,
                    var_dim='chrom5k')

atac_ds

adata = atac_ds.get_count_adata(da_name='chrom5k_da',
                                binarize_cutoff=0,
                                sparse=True,
                                loading_chunk=30000)

## Basic Feature Filtering

chroms = [
    'chr1', 'chr2', 'chr3', 'chr4', 'chr5', 'chr6', 'chr7', 'chr8', 'chr9',
    'chr10', 'chr11', 'chr12', 'chr13', 'chr14', 'chr15', 'chr16', 'chr17',
    'chr18', 'chr19'
]

remove_chromosomes(adata, include_chromosomes=chroms)

filter_regions(adata, n_cell=n_cell, zscore_abs_cutoff=zscore_abs_cutoff)

remove_black_list_region(adata, black_list_path=mm10.ENCODE_BLACKLIST_PATH)

adata.write_h5ad('atac_input.h5ad')

pd.Series(adata.var_names).to_csv('atac_features.txt', index=None, header=False)

adata
